requirements.txt
# Core
numpy==1.26.4
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10

# --- HuggingFace (PINNED for CI stability) ---
transformers==4.41.2
huggingface_hub==0.23.4
tokenizers==0.19.1
accelerate==0.31.0
datasets==2.19.1
evaluate==0.4.2
safetensors==0.4.3
sentencepiece==0.2.0
torch==2.3.1

# Metrics & utilities
tqdm>=4.66
regex>=2023.12.25

# Experiment tracking (MLflow)
mlflow>=2.10
sqlalchemy>=2.0
psutil>=5.9

# DVC (data & model versioning)
dvc>=3.0

# Config / pipeline params
pyyaml>=6.0
hydra-core>=1.3

# Notebooks (optional for CI, but OK to keep)
jupyter>=1.0
ipykernel>=6.29

# Plotting / analysis
matplotlib>=3.7
seaborn>=0.13

# Serving (repo has app/api.py)
flask>=2.3
gunicorn>=21.2

# Testing
pytest>=7.4
pytest-cov>=4.1
requests>=2.31

# Code quality / CI tools
black>=24.1
isort>=5.13
flake8>=7.0

# Packaging helpers
python-dotenv>=1.0

pyarrow==15.0.2



